requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
schedule==1.2.1
loguru==0.7.2
pymongo==4.6.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-core
email-validator==2.1.0.post1
requests==2.31.0
psycopg2